[Shortcode] Extract functions that possibly can be used for widget as well
